clean orphaned notifcations3 (#20068)
* handle common reports

* dont remove while iterating

* that wouldve been bad

+++ corehq/apps/reports/management/commands/clean_orphaned_notifications.py
@@ -3,6 +3,7 @@ from __future__ import unicode_literals
 from django.core.management import BaseCommand
 from corehq.apps.reports.models import ReportNotification, ReportConfig
 from corehq.apps.userreports.models import ReportConfiguration
+from corehq.util.couch import get_document_or_not_found, DocumentNotFound
 from collections import defaultdict
 from couchdbkit.exceptions import ResourceNotFound
 from dimagi.utils.couch.undo import is_deleted
@@ -30,26 +31,32 @@ class Command(BaseCommand):
         for notification_id in notification_id_iterator:
             nid = notification_id['id']
             notification = ReportNotification.get(nid)
-            updated_notification = False
+            cids_to_remove = []
 
             for cid in notification.config_ids:
                 if not existent_config_ids[cid]:
                     try:
                         rc = ReportConfig.get(cid)
-                        rcuration = ReportConfiguration.get(rc.subreport_slug)
+                        if not rc.subreport_slug:  # seems to be the case of common reports
+                            existent_config_ids[cid] = True
+                            continue
+
+                        rcuration = get_document_or_not_found(
+                            ReportConfiguration, rc.domain, rc.subreport_slug)
                         if is_deleted(rcuration):
                             if options['execute']:
                                 super(type(rc), rc).delete()  # i am updating notifications manually
-                            notification.config_ids.remove(cid)
-                            updated_notification = True
-                        else:
-                            existent_config_ids[cid] = True
-                    except ResourceNotFound:
-                        notification.config_ids.remove(cid)
-                        updated_notification = True
+                            cids_to_remove.append(cid)
+                    except ResourceNotFound:  # ReportConfig not found
+                        cids_to_remove.append(cid)
+                    except DocumentNotFound:  # ReportConfiguration not found
+                        pass
 
-            if updated_notification:
+            if cids_to_remove:
                 if options['execute']:
+                    for cid in cids_to_remove:
+                        notification.config_ids.remove(cid)
+
                     if notification.config_ids:
                         notification.save()
                     else:
